home nav bug fix
